add the react-icons package and refactor header code
