+++ src/Components/NavBar/BlogButton.js
@@ -1,0 +1,14 @@
+function BlogButton() {
+  return (
+    <a
+      className="navItem"
+      href="https://jackparsonss.hashnode.dev/"
+      target="_blank"
+      rel="noopener noreferrer"
+    >
+      <h3>Blog</h3>
+    </a>
+  );
+}
+
+export default BlogButton;
