+++ src/Components/NavBar/Header.css
@@ -2,6 +2,12 @@
   font-family: "Fira Code", monospace;
   height: 7.5%;
   background-color: #22262b;
+  background: linear-gradient(
+    90deg,
+    rgba(18, 19, 20, 1) 0%,
+    rgba(34, 38, 43, 1) 59%,
+    rgba(49, 54, 59, 1) 100%
+  );
   color: #fff;
   display: grid;
   grid-template-areas: "logo nav";
@@ -10,7 +16,7 @@
 .header__nav {
   grid-area: nav;
   display: grid;
-  grid-template-columns: repeat(6, auto);
+  grid-template-columns: repeat(5, auto);
   align-items: center;
   justify-items: center;
 }
