+++ src/Components/NavBar/NavItems.js
@@ -1,0 +1,14 @@
+export const navItems = [
+  {
+    name: "About",
+    link: "about",
+  },
+  {
+    name: "Projects",
+    link: "project1",
+  },
+  {
+    name: "Contact",
+    link: "contact",
+  },
+];
