+++ src/Components/NavBar/ResumeButton.css
@@ -1,0 +1,10 @@
+.resumeButton {
+  color: #5240fe;
+  border: 2px solid #5240fe;
+  padding: 5px 10px;
+  border-radius: 10px;
+}
+
+.resumeButton:hover {
+  color: white;
+}
