+++ src/Components/NavBar/ResumeButton.js
@@ -1,0 +1,16 @@
+import { Link } from "react-router-dom";
+import "./ResumeButton.css";
+
+function ResumeButton() {
+  return (
+    <Link
+      className="navItem resumeButton"
+      to="/jack-parsons-resume.pdf"
+      target="_blank"
+    >
+      <h3>Resume</h3>
+    </Link>
+  );
+}
+
+export default ResumeButton;
